+++ app/controllers/posts_controller.php
@@ -490,14 +490,21 @@ class PostsController extends AppController {
 			$this->Session->setFlash(__('Invalid id for post', true));
 			$this->redirect($this->referer());
 		}
-		// second param = cascade -> delete associated records from hasmany , hasone relations
-		if ($this->Post->delete($id, true)) {
-			$this->Session->setFlash(__('Post deleted', true), 'default', array('class' => 'success'));
-			$this->redirect(array('controller' => 'users',  'action' => 'view',  $this->Session->read('Auth.User.id')));
-		}
-		$this->Session->setFlash(__('Post was not deleted', true));
-		$this->redirect($this->referer());
-	}
+        $this->Post->contain();
+        $post =  $this->Post->read(array('id','user_id'), $id);
+        if($post['Post']['user_id'] == $this->Session->read('Auth.User.id')){
+            // second param = cascade -> delete associated records from hasmany , hasone relations
+            if ($this->Post->delete($id, true)) {
+                $this->Session->setFlash(__('Post deleted', true), 'default', array('class' => 'success'));
+                $this->redirect(array('controller' => 'users',  'action' => 'view',  $this->Session->read('Auth.User.id')));
+            }
+            $this->Session->setFlash(__('Post was not deleted', true));
+            $this->redirect($this->referer());
+        } else {
+            $this->Session->setFlash(__('The Post does not belong to you.', true));
+            $this->redirect($this->referer());
+        }
+    }
 
 	/**
 	 * action to be called from multiple file upload for add / edit
